added an option to setup a local syncd-rpc for QoS SAI tests (#5338)
What is the motivation for this PR?
QoS SAI tests require syncd-rpc to be installed. The syncd-rpc image is expected to be located at Docker registry where it is pulled from.

This PR brings an alternative for syncd-rpc setup for the case when its image is present locally in SONiC. If the local image is present it's favored and will be installed instead of looking on registry.