[qos] Disable FDB aging before start of test (#5452)
Few qos testcases were consistently failing on backend T0 topology. This test populates the fdb prior to the start of the test run. The failures were due to mac entries aging in the middle of the tests and hence the tests not seeing the expected number of packets in the respective pg/queue of the desired port.

This PR contains the following changes
* Disable fdb aging before populating fdb entries and then restore the default aging value at the end of all the tests
* add more logging to the PGShared and QShared watermark tests to get the number of packets received. This will help rule out slow PTF possibility

How did you verify/test it?
Reproduced the problem on backend T0. Reran the testcases with the changes in this PR and could no longer see the issue.